requirements.txt
Django==1.11.6
Pillow==4.3.0
psycopg2==2.7.3.1
djangorestframework
drfdocs